telegram/requirements.txt
aiogram==2.25.1
python-dotenv==0.21.1
aiohttp==3.8.4
aioschedule==0.5.2

